+++ CHANGELOG.md
@@ -8,3 +8,5 @@ Nothing yet.
 
 * Added an interface that all Guarantees must support.
   * Added `GanbaroDigital\Guarantees\Values\Guarantee`
+* Added support for wrapping Traversable data.
+  * Added `GanbaroDigital\Guarantees\Values\TraversableGuarantee`